### 2022-03-19 #4254 @editorialbot
[PRE REVIEW]: BetterReg: An R package for Useful Regression Statistics
**Submitting author:** @chrisaberson (<a href="http://orcid.org/0000-0003-3481-7177">Christopher Aberson</a>)
**Repository:** https://github.com/chrisaberson/BetterReg
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
