### 2022-03-21 #4254 @chrisaberson
@danielkatz

I have updated the licence. My apologies. I was unaware of the differences
between licences. I just used what I saw others using. Licence has been
changed to GNU public 3

On Mon, Mar 21, 2022 at 8:31 AM Daniel S. Katz ***@***.***>
wrote:

> 👋 @chrisaberson <https://github.com/chrisaberson> - I think work is
> needed in your README to meet JOSS criteria, but we could wait for this to
> come up in the review process.
>
> A more serious issue, however, is that the license you have chosen (CC0)
> is not an OSI-approved license, so JOSS cannot review/publish this work.
> See https://opensource.org/faq#cc-zero. If you are willing to change the
> license to an OSI-approved license, we can start the review process. If
> not, we will have to mark this as rejected.
>
