### 2022-03-27 #4254 @jbytecode
👋👋👋 Dear @62442katieb  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**BetterReg: An R package for Useful Regression Statistics**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4254).
