### 2022-03-28 #4254 @sestelo
> 👋👋👋 Dear @sestelo 👋👋👋
>
> Would you be willing to assist in reviewing this submission for **JOSS** (**Journal of Open Source Software**)?
>
> JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**BetterReg: An R package for Useful Regression Statistics**". You can find more information at the top of this Github issue (#4254).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.

Dear @jbytecode,

I'm sorry but, at this moment, I can't review this submission... :( best!
