+++ R/R2change.Rd
@@ -1,21 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/R2change.R
-\name{R2change}
-\alias{R2change}
-\title{R-square change for Hierarchical Multiple Regression}
-\usage{
-R2change(model1 = NULL, model2 = NULL)
-}
-\arguments{
-\item{model1}{first regression model}
-
-\item{model2}{second regression model}
-}
-\description{
-R-square change for Hierarchical Multiple Regression
-}
-\examples{
-mymodel1<-lm(y~x1+x2, data=testreg)
-mymodel2<-lm(y~x1+x2+x3+x4, data=testreg)
-R2change(model1=mymodel1, model2=mymodel2)
-}
