### 2022-04-16 #4280 @brunomontezano
> Dear reviewers @62442katieb and @brunomontezano
>
> How is your review going? Could you please update your status?
>
> Thank you in advance!

Dear editor @jbytecode,

I've been gathering notes and creating issues on the authors repository to improve the quality of his submission. I'm aiming to finish the process soon.

This last week was pretty busy for me. But thanks for getting in touch! I remain at your disposal for further demands.

Sincerely,

Bruno Montezano
