### 2022-04-16 #4280 @chrisaberson
I believe this was resolved. There was a corrupt file in the R directory. Fixed. Also, CRAN shows all checks being passed https://cran.r-project.org/web/checks/check_results_BetterReg.html

I just installed using devtools::install_github("chrisaberson/BetterReg", force=TRUE) and everything went fine
