+++ README.md
@@ -33,35 +33,35 @@ This project is licensed under GNU General Public License version 3.
 
 ### part function for squared semipartial correlations  
 
-The part function requires an existing LM model and indication of number of predictors.
+The part function requires an existing LM model and indication of number of predictors.  
 
-library(BetterReg)
-mymodel<-lm(y~x1+x2+x3+x4+x5, data=testreg)
-parts(model=mymodel, pred=5)
+library(BetterReg)  
+mymodel<-lm(y~x1+x2+x3+x4+x5, data=testreg)  
+parts(model=mymodel, pred=5)  
 
- Predictor 1: semi partial = 0.032; squared semipartial = 0.001
- Predictor 2: semi partial = 0.307; squared semipartial = 0.094
- Predictor 3: semi partial = 0.268; squared semipartial = 0.072
- Predictor 4: semi partial = 0.134; squared semipartial = 0.018
- Predictor 5: semi partial = 0.241; squared semipartial = 0.058
+ Predictor 1: semi partial = 0.032; squared semipartial = 0.001  
+ Predictor 2: semi partial = 0.307; squared semipartial = 0.094  
+ Predictor 3: semi partial = 0.268; squared semipartial = 0.072  
+ Predictor 4: semi partial = 0.134; squared semipartial = 0.018  
+ Predictor 5: semi partial = 0.241; squared semipartial = 0.058  
 
 ### tolerance function for multicollinearity assumptions  
 
-The tolerance function requires only a model.
+The tolerance function requires only a model.  
 
-mymodel<-lm(y~x1+x2+x3+x4+x5, data=testreg)
-tolerance(model=mymodel)
-
+mymodel<-lm(y~x1+x2+x3+x4+x5, data=testreg)  
+tolerance(model=mymodel)  
+ 
         x1        x2        x3        x4        x5 
  0.9976977 0.9990479 0.9931082 0.9953317 0.9980628
 
 
 ### Mahal function for detecting multivariate outliers  
 
-The Mahal function requires model, predictors, and desired number of values to output.
+The Mahal function requires model, predictors, and desired number of values to output.  
 
-mymodel<-lm(y~x1+x2+x3+x4+x5, data=testreg)
-Mahal(model=mymodel, pred=5, values=10)
+mymodel<-lm(y~x1+x2+x3+x4+x5, data=testreg)  
+Mahal(model=mymodel, pred=5, values=10)  
 
       537      770      342      760      299      982      446      174 
  14.56342 15.03188 15.56224 15.60986 16.52869 16.80958 17.38597 18.11072 
@@ -71,20 +71,20 @@ Mahal(model=mymodel, pred=5, values=10)
 
 ### LRchi function for Logistic Regression Coefficients  
 
-The LRchi function takes input for the dependent variable name (y), up to 10 predictors (x1, x2, etc.), and the number of predictors.
+The LRchi function takes input for the dependent variable name (y), up to 10 predictors (x1, x2, etc.), and the number of predictors.  
 
-LRchi(data=testlog, y="dv", x1="iv1", x2="iv2",numpred=2)
+LRchi(data=testlog, y="dv", x1="iv1", x2="iv2",numpred=2)  
 
-Predictor: iv1; LR squared 34.09, p= 0
-Predictor: iv2; LR squared 0.19, p= 0.67
+Predictor: iv1; LR squared 34.09, p= 0  
+Predictor: iv2; LR squared 0.19, p= 0.67  
 
 ### Pseudo function for Logistic Regression Effect Size  
 
-The Psuedo function takes an existing model as input
[…]
