### 2022-05-06 #4280 @jbytecode
@chrisaberson - First you can start with correcting the suggestions of @62442katieb and please report here your changes, mention your commits so we can go on with the further suggestions
