### 2022-05-06 6f76da2
added more detailed examples
