### 2022-05-10 #4280 @chrisaberson
I have added a statement of need, how R users currently calculate these measures, and the target audience. OLS is now written out. I have added a detailed real world example to the repositories readme page and addressed community guidelines
