### 2022-05-11 4f84ed3
Fixing minor error in b comparisons
