+++ R/indbcomp.R
@@ -16,25 +16,18 @@
 #'
 
 
-
-indbcomp<-function(model1=NULL, model2=NULL, comps="abs")
+indbcomp<-function(model1=NULL, model2=NULL, comps="abs", pred=NULL)
 {
+  if ((pred =="2") && (comps=="abs")){
 
-  pred1<-model1$df[1]-1
-  pred2<-model2$df[1]-1
-
-  try(if(pred1!=pred2) stop("Models must be identical"))
-
-  if ((pred1 =="2") && (comps=="abs")){
-
-  b1_1<-(model1$coefficients)[2,1]
-  b2_1<-(model1$coefficients)[3,1]
-  b1_2<-(model2$coefficients)[2,1]
-  b2_2<-(model2$coefficients)[3,1]
-  seb1_1<-(model1$coefficients)[2,2]
-  seb2_1<-(model1$coefficients)[3,2]
-  seb1_2<-(model2$coefficients)[2,2]
-  seb2_2<-(model2$coefficients)[3,2]
+  b1_1<-(model1$coefficients)[2,4]
+  b2_1<-(model1$coefficients)[3,4]
+  b1_2<-(model2$coefficients)[2,4]
+  b2_2<-(model2$coefficients)[3,4]
+  seb1_1<-model1$coefficients[2,2]
+  seb2_1<-model1$coefficients[3,2]
+  seb1_2<-model2$coefficients[2,2]
+  seb2_2<-model2$coefficients[3,2]
   sebb1<-((seb1_1^2)+(seb2_1^2))^.5
   sebb2<-((seb1_2^2)+(seb2_2^2))^.5
   t1<-abs(abs(b1_1)-abs(b1_2))/sebb1
@@ -50,20 +43,20 @@ indbcomp<-function(model1=NULL, model2=NULL, comps="abs")
   }
 
 
-  if ((pred1 =="3") && (comps=="abs")){
+  if ((pred =="3") && (comps=="abs")){
 
-    b1_1<-(model1$coefficients)[2,1]
-    b2_1<-(model1$coefficients)[3,1]
-    b3_1<-(model1$coefficients)[4,1]
-    b1_2<-(model2$coefficients)[2,1]
-    b2_2<-(model2$coefficients)[3,1]
-    b3_2<-(model2$coefficients)[4,1]
-    seb1_1<-(model1$coefficients)[2,2]
-    seb2_1<-(model1$coefficients)[3,2]
-    seb3_1<-(model1$coefficients)[4,2]
-    seb1_2<-(model2$coefficients)[2,2]
-    seb2_2<-(model2$coefficients)[3,2]
-    seb3_2<-(model2$coefficients)[4,2]
+    b1_1<-(model1$coefficients)[2,4]
+    b2_1<-(model1$coefficients)[3,4]
+    b3_1<-(model1$coefficients)[4,4]
+    b1_2<-(model2$coefficients)[2,4]
+    b2_2<-(model2$coefficients)[3,4]
+    b3_2<-(model2$coefficients)[4,4]
+    seb1_1<-model1$coefficients[2,2]
+    seb2_1<-model1$coefficients[3,2]
+    seb3_1<-model1$coefficients[4,2]
+    seb1_2<-model2$coefficients[2,2]
+    seb2_2<-model2$coefficients[3,2]
+    seb3_2<-model2$coefficients[4,2]
     sebb1<-((seb1_1^2)+(seb1_2^2))^.5
     sebb2<-((seb2_1^2)+(seb2_2^2))^.5
     sebb3<-((seb3_1^2)+(seb3_2^2))^.5
@@ -83,24 +76,24 @@ indbcomp<-function(model1=NULL, model2=NULL, comps="abs")
   }
 
 
-  if ((pred1 =="4") && (comps=="abs")){
[…]
