+++ R/depbcomp.R
@@ -11,7 +11,7 @@
 #'@param numpred number of predictors
 #'@param comps Type of comparison, "abs" for absolute values or "raw" for raw coefficients
 #'@examples
-#'depbcomp(data=testreg,y=y,x1=x1,x2=x2,x3=x3,x4=x4,x5=x5, numpred=5,comps="abs")
+#'depbcomp(data=testreg,y="y",x1="x1",x2="x2",x3="x3",x4="x4",x5="x5", numpred=5,comps="abs")
 #'@return Comparing Dependent Coefficients in Multiple Regression
 #'@export
 #'
