+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: BetterReg
 Type: Package
 Title: Better Statistics for OLS and Binomial Logistic Regression
-Version: 0.2.0
+Version: 0.3.0
 Author: Chris Aberson
 Maintainer: Chris Aberson <cla18@humboldt.edu> 
 Description: Provides squared semi partial correlations, tolerance, Mahalanobis, Likelihood Ratio Chi Square, and Pseudo R Square. Aberson, C. L. (2022) <doi:10.31234/osf.io/s2yqn>.
