+++ README.md
@@ -1,6 +1,6 @@
 # BetterReg Readme
 
-Chris Aberson May 2, 2022
+Chris Aberson May 13, 2022
 
 # BetterReg  
 
