### 2022-05-25 1ded21d
Correct minor misspelling and function names
