+++ README.md
@@ -35,17 +35,17 @@ Using data from Aberson (2007), the analyses that follow predict support for Aff
 
 The `part` function requires an existing LM model and indication of number of predictors.  
 
-First, build the model:
-xx<-lm(formula = AA_DV ~ AA_Initial+pers_exp+liberal+female+economic, data = hand5)
+First, build the model:  
+xx<-lm(formula = AA_DV ~ AA_Initial+pers_exp+liberal+female+economic, data = hand5)  
 
-Then, using the parts command, provide the model name and the number of predictors
-parts(model=xx, pred=5)
+Then, using the parts command, provide the model name and the number of predictors  
+parts(model=xx, pred=5)  
 
-Predictor 1: semi partial = 0.333; squared semipartial = 0.111 
-Predictor 2: semi partial = 0.032; squared semipartial = 0.001 
-Predictor 3: semi partial = 0.197; squared semipartial = 0.039 
-Predictor 4: semi partial = 0.095; squared semipartial = 0.009 
-Predictor 5: semi partial = 0.032; squared semipartial = 0.001 
+Predictor 1: semi partial = 0.333; squared semipartial = 0.111  
+Predictor 2: semi partial = 0.032; squared semipartial = 0.001  
+Predictor 3: semi partial = 0.197; squared semipartial = 0.039  
+Predictor 4: semi partial = 0.095; squared semipartial = 0.009  
+Predictor 5: semi partial = 0.032; squared semipartial = 0.001  
 
 The `Mahal` function provides Mahalanobis values and requires input of model and predictors as well as the number of values to return (10 is the default). 
 
@@ -59,10 +59,9 @@ The `tolerance` command requires only the model name.
 
 tolerance(model=xx)  
 AA_Initial   pers_exp    liberal     female   economic  
-
 0.9464682  0.9904156  0.9058256  0.9418196  0.9910559   
 
-`R2change` compares two models. Below, I added merit, discrimination, and diversity participation to the model (xx2).
+`R2change` compares two models. Below, I added merit, discrimination, and diversity participation to the model (xx2).  
 the R2change command takes model1 (xx) and compares it to model2 (xx2). Note that this approach is only for models that are adding variables to a previous model.
 
 xx2<-lm(formula = AA_DV ~ AA_Initial+pers_exp+liberal+female+economic+merit+discrim+div_part, data = hand5) 
@@ -72,18 +71,18 @@ R-square change = 0.181
 
 F(3,704) = 70.537, p = 6.81538788796511e-40 
 
-The `depbcomp` function allows for comparisons of  dependent coefficients. These are coefficients in the same model. 
+The `depbcomp` function allows for comparisons of  dependent coefficients. These are coefficients in the same model.   
 
-depbcomp(data=sample1,y="AA_DV",x1="div_part",x2="merit", x3="discrim",numpred=3,comps="abs")
-Pred 1 vs. Pred 2  : t = 4.633, p = 4.28730081880602e-06
-Pred 1 vs. Pred 3  : t = 9.614, p = 0
-Pred 2 vs. Pred 3  : t = 5.371, p = 1.0627416191511e-07
+depbcomp(data=sample1,y="AA_DV",x1="div_part",x2="merit", x3="discrim",numpred=3,comps="abs")  
+Pred 1 vs. Pred 2  : t = 4.633, p = 4.28730081880602e-06  
+Pred 1 vs. Pred 3  : t = 9.614, p = 0  
+Pred 2 vs. Pred 3  : t = 5.371, p = 1.0627416191511e-07  
 
[…]
