+++ JOSS/paper.bib
@@ -4,12 +4,11 @@
 	title = {Logistic regression: from introductory to advanced concepts and applications},
 	isbn = {978-1-4129-7483-7 978-1-4129-7119-5},
 	shorttitle = {Logistic regression},
+	url = {https://us.sagepub.com/en-us/nam/logistic-regression/book227554},
 	publisher = {SAGE},
 	author = {Menard, Scott W.},
 	year = {2010},
-	note = {OCLC: ocn277118375},
 	keywords = {Logistic distribution, Logistic regression analysis},
-	annote = {Introduction : linear regression and logistic regression -- Log-linear analysis, logit analysis, and logistic regression -- Quantitative approaches to model fit and explained variation -- Prediction tables and qualitative approaches to explained variation -- Logistic regression coefficients -- Model specification, variable selection, and model building -- Logistic regression diagnostics and problems of inference -- Path analysis with logistic regression (PALR) -- Polytomous logistic regression for unordered categorical variables -- Ordinal logistic regression -- Clusters, contexts, and dependent data : logistic regression for clustered sample survey data -- Conditional logistic regression models for related samples -- Longitudinal panel analysis with logistic regression -- Logistic regression for historical and developmental change models : multilevel logistic regression and discrete time event history analysis -- Comparisons : logistic regression and alternative models},
 }
 
 @book{tabachnick_using_2019,
@@ -17,31 +16,22 @@
 	edition = {Seventh edition},
 	title = {Using multivariate statistics},
 	isbn = {978-0-13-479054-1},
+	url = {https://www.pearson.com/us/higher-education/program/Tabachnick-Using-Multivariate-Statistics-7th-Edition/PGM2458367.html},
 	publisher = {Pearson},
 	author = {Tabachnick, Barbara G. and Fidell, Linda S. and Ullman, Jodie B.},
 	year = {2019},
 	keywords = {Statistics, Multivariate analysis},
-	annote = {Chapter 14, by Jodie B. Ullman},
-	annote = {Introduction -- A guide to statistical techniques : using the book -- Review of univariate and bivariate statistics -- Cleaning up your act : screening data -- Multiple regression -- Analysis of covariance -- Multivariate analysis of variance and covariance -- Profile analysis : the multivariate approach to repeated measures -- Discriminant analysis -- Logistic regression -- Survival-failure analysis -- Canonical correlation -- Principal components and factor analysis -- Structural equation modeling / by Jodie B. Ullman -- Multilevel linear modeling -- Multiway frequency analysis -- An overview of the general linear model -- Time-series analysis (available online)},
 }
-  @Book{fox_r_2019,
-    title = {An {R} Companion to Applied Regression},
-    edition = {Third},
-    author = {John Fox and Sanford Weisberg},
-    year = {2019},
-    publisher = {Sage},
-    address = {Thousand Oaks {CA}},
-    url = {https://socialsciences.mcmaster.ca/jfox/Books/Companion/},
-  }
 
-@Manual{R_core_team,
-    title = {R: A Language and Environment for Statistical Computing},
-    author = {{R Core Team}},
-    organization = {R Foundation for Statistical Computing},
-    address = {Vienna, Austria},
-    year = {2021},
-    url = {https://www.R-project.org/},
-  }
+@book{fox_r_2019,
+	address = {Thousand Oaks CA},
+	edition = {Third},
+	title = {An {R} {Companion} to {Applied} {Regression}},
+	url = {https://socialsciences.mcmaster.ca/jfox/Books/Companion/},
+	publisher = {Sage},
[…]
