### 2022-05-29 chrisaberson/BetterReg#8 @brunomontezano
Final demands
I believe that most of my suggestions were accepted. I think what still has room for improvement would be:

- [x] Improved final proof review regarding citations (some are still being compiling weirdly, and you can check it on the proof at the review issue) - pay attention to the following section of the documentation (https://joss.readthedocs.io/en/latest/submitting.html?highlight=citation#example-paper-and-bibliography)

![Screenshot2022-05-2913:55:31](https://user-images.githubusercontent.com/65104127/170882163-cc071ec5-198f-404b-abd3-1809fecf6e97.png)

- [x] Final review of code syntax in the README of the repository. There are inconsistencies (code chunks without code syntax). The commands must be in `code syntax`, and output as well (I suggest inserting two hashes [##] before the texts that are output). I leave some examples below, but I suggest using this standard in the whole README.

![Screenshot2022-05-2913:58:57](https://user-images.githubusercontent.com/65104127/170882359-0e87629a-36f5-42f9-83f8-d9e6d3807c16.png)
![Screenshot2022-05-2913:59:12](https://user-images.githubusercontent.com/65104127/170882368-1d6426a1-33dd-4564-8b4d-1e00306406c5.png)

- [x] I'd add the `install.packages("BetterReg")` in the installation instructions in the README as well, since the package is available on CRAN. It's a super positive aspect of your work.

In addition, I think the work is very well suited and of sufficient quality for publication.

Edit to point out the relevant part of the documentation on a quick reference of citations:

For a quick reference, the following citation commands can be used:
- `@author:2001`  ->  "Author et al. (2001)"
- `[@author:2001]` -> "(Author et al., 2001)"
- `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
