### 2022-05-29 chrisaberson/BetterReg#8 @brunomontezano
Due to https://github.com/openjournals/joss-reviews/issues/4280
