+++ JOSS/paper.bib
@@ -1,52 +1,41 @@
 
 @book{menard_logistic_2010,
-	address = {Los Angeles},
 	title = {Logistic regression: from introductory to advanced concepts and applications},
-	isbn = {978-1-4129-7483-7 978-1-4129-7119-5},
-	shorttitle = {Logistic regression},
+	publisher = {Sage. Thousand Oaks, U.S.A},
 	url = {https://us.sagepub.com/en-us/nam/logistic-regression/book227554},
-	publisher = {SAGE},
-	author = {Menard, Scott W.},
-	year = {2010},
-	keywords = {Logistic distribution, Logistic regression analysis},
+	author = {Menard, S. W.},
+	year = 2010
 }
 
 @book{tabachnick_using_2019,
-	address = {NY, NY},
-	edition = {Seventh edition},
-	title = {Using multivariate statistics},
-	isbn = {978-0-13-479054-1},
+	title = {Using multivariate statistics (Seventh edition)},
+	Publisher = {Pearson. London, UK},
 	url = {https://www.pearson.com/us/higher-education/program/Tabachnick-Using-Multivariate-Statistics-7th-Edition/PGM2458367.html},
-	publisher = {Pearson},
-	author = {Tabachnick, Barbara G. and Fidell, Linda S. and Ullman, Jodie B.},
+	author = {{Tabachnick}, B. G. and {Fidell}, L. S. and {Ullman}, J. B.},
 	year = {2019},
-	keywords = {Statistics, Multivariate analysis},
 }
 
 @book{fox_r_2019,
-	address = {Thousand Oaks CA},
-	edition = {Third},
-	title = {An {R} {Companion} to {Applied} {Regression}},
+	title = {An R Companion to Applied Regression (Third edition)},
+	publisher = {Sage. Thousand Oaks, U.S.A},
 	url = {https://socialsciences.mcmaster.ca/jfox/Books/Companion/},
-	publisher = {Sage},
-	author = {Fox, John and Weisberg, Sanford},
+	author = {{Fox}, J. and {Weisberg}, S.},
 	year = {2019},
 }
 
 @book{cohen_applied_2015,
-	title = {Applied multiple regression/correlation analysis for the behavioral sciences},
-	isbn = {978-0-203-77444-1 978-1-134-80101-5 978-1-4106-0626-6 978-1-134-80094-0},
+	title = {Applied multiple regression/correlation analysis for the behavioral sciences.},
+	publisher = {Routledge. London, UK},
 	url = {https://www.taylorfrancis.com/books/e/9780203774441},
-	abstract = {This classic text on multiple regression is noted for its nonmathematical, applied, and data-analytic approach. Readers profit from its verbal-conceptual exposition and frequent use of examples. The applied emphasis provides clear illustrations of the principles and provides worked examples of the types of applications that are possible. Researchers learn how to specify regression models that directly address their research questions. An overview of the fundamental ideas of multiple regression and a review of bivariate correlation and regression and other elementary statistical concepts provide a strong foundation for understanding the rest of the text. The third edition features an increased emphasis on graphics and the use of confidence intervals and effect size measures, and an accompanying website with data for most of the numerical examples along with the computer code for SPSS, SAS, and SYSTAT, at www.psypress.com/9780805822236 . Applied Multiple Regression serves as both a textbook for graduate students and as a reference tool for researchers in psychology, education, health sciences, communications, business, sociology, political science, anthropology, and economics. An introductory knowledge of statistics is required. Self-standing chapters minimize the need for researchers to refer to previous chapters.},
 	language = {English},
 	urldate = {2022-02-21},
[…]
