### 2022-05-30 chrisaberson/BetterReg#8 @chrisaberson
thank you for the detailed comments and direction. I am admittedly a novice using .bib format, so the extra guidance was much appreciated.

I believe I have addressed all issues but am happy to make additional changes.

I'm finally covid negative and made it through largely symptom free.
