### 2022-06-01 2902c40
add missing doi and correct type in bibtex
