+++ JOSS/paper.bib
@@ -16,7 +16,7 @@
 }
 
 @book{fox_r_2019,
-	title = {An R Companion to Applied Regression (Third edition)},
+	title = {An {R} Companion to Applied Regression (Third edition)},
 	publisher = {Sage. Thousand Oaks, U.S.A},
 	url = {https://socialsciences.mcmaster.ca/jfox/Books/Companion/},
 	author = {{Fox}, J. and {Weisberg}, S.},
@@ -31,11 +31,12 @@
 	urldate = {2022-02-21},
 	author = {{Cohen}, J. and {Cohen}, P. and {West} and S. G. and {Aiken}, L. S.},
 	year = {2015},
+	doi  = {10.4324/9780203774441},
 }
 
 @book{r_core_team_r_2022,
 	address = {Vienna, Austria},
-	title = {R: A Language and Environment for Statistical Computing},
+	title = {{R}: A Language and Environment for Statistical Computing},
 	url = {https://www.R-project.org/},
 	publisher = {R Foundation for Statistical Computing},
 	author = {{R Core Team}},
