+++ JOSS/paper.bib
@@ -23,14 +23,14 @@
 	year = {2019},
 }
 
-@book{cohen_applied_2015,
+@book{cohen_applied_2002,
 	title = {Applied multiple regression/correlation analysis for the behavioral sciences.},
 	publisher = {Routledge. London, UK},
 	url = {https://www.taylorfrancis.com/books/e/9780203774441},
 	language = {English},
 	urldate = {2022-02-21},
 	author = {{Cohen}, J. and {Cohen}, P. and {West} and S. G. and {Aiken}, L. S.},
-	year = {2015},
+	year = {2002},
 }
 
 @book{r_core_team_r_2022,
