+++ JOSS/paper.md
@@ -18,25 +18,24 @@ bibliography: paper.bib
 
 # Summary
 
-Statistics such as squared semi partial correlations, tolerance, and Mahalanobis Distances are useful for reporting the results of OLS Regression [@tabachnick_using_2019]. Similarly values such as the Likelihood Ratio Chi-square  [@cohen_applied_2015] and Likelihood R-square [@menard_logistic_2010]. Such statistics are not part of base R [@r_core_team_r_2022] popular packages such *car* [@fox_r_2019]. To fill these gaps, I developed BetterReg to provide these values. 
+Statistics such as squared semi partial correlations, tolerance, and Mahalanobis Distances are useful for reporting the results of OLS (Ordinary Least Squares) Regression [@tabachnick_using_2019]. Similarly values such as the Likelihood Ratio Chi-square  [@cohen_applied_2015] and Likelihood R-square [@menard_logistic_2010]. Such statistics are not part of base R [@r_core_team_r_2022] popular packages such as *car* [@fox_r_2019]. To fill these gaps, the BetterReg package is developed to provide these statistics and measures. 
 
-Squared semipartial correlations provide a measure of uniquely explained variances that is on the same scale as $R^2$ values. Tolerance values address multicollinearity by addressing variance unexplained in a predictor. Mahalanabis Distances are a popular measure of multivariate outliers that are presented on an $\chi^2$ scale. The Likelihood Ratio $\chi^2$ provides a significance test that is more stable than the commonly presented Wald Test for logistic regression. The Likelihood Ratio $\chi^2$ is the most widely recommended Pseudo $R^2$ statistic for logistic. 
+Squared semipartial correlations provide a measure of uniquely explained variances that is on the same scale as $R^2$ values. Tolerance values address multi-collinearity by addressing variance unexplained in a predictor. Mahalanabis Distance is a popular measure of multi-variate outliers that are presented on a $\chi^2$ scale. The Likelihood Ratio $\chi^2$ provides a significance test that is more stable than the commonly presented Wald Test and the Likelihood Ratio $\chi^2$ is the most widely recommended Pseudo $R^2$ statistic for the Logistic Regression. 
 
-The target audience for this package is researchers using Ordinary Least Squares (OLS) and Logistic Regression. Presently, I know of no R packages that provide these statistics, so calculation requires researchers to write their own code. These values are widely available in commercial programs such as SAS, SPSS, and Stata.
+The target audience for this package is researchers using OLS and Logistic Regression. Presently, there is not any R packages that provide these statistics, so the calculation requires researchers to write their own code. These statistics are widely available in commercial programs such as SAS, SPSS, and Stata.
 
 # Usage
 
-BetterReg functions require existing regression models (either OLS or Logistic for most statistics), dataset names (for some approaches), number of predictors (some functions), and desired amount of output (Mahal function). 
+BetterReg functions require existing regression models (either OLS or Logistic for most statistics), dataset names (for some approaches), number of predictors (some functions), and desired amount of output (the Mahal function). 
 
 ## `part` function for squared semipartial correlations
 
 The `part` function requires an existing LM model and indication of
-number of predictors.
[…]
