### 2022-06-01 #4280 @jbytecode
@chrisaberson - we have still some rooms for possible improvements. please check the pull request.
