### 2022-06-02 #4280 @jbytecode
@chrisaberson - we have one more issue about the paper (code coloring for R code). Please merge the PR.
